rc: network.subr improve network6_getladr()
In network6_getladdr() we are iterating over inet6 lines and are not
interested in any others.  So tell ifconfig to limit output to "inet6"
as much as possible.
This is probably a micro-optimisation but was noticed while looking
at other IPv6-related boot-time improvements.

Taken from: FreeBSD (https://github.com/freebsd/freebsd-src/commit/e448ff6f0)