Initial port of uvc driver from FreeBSD
The port is based on [1] (branch feature-uvc-prepare-submit3) as of
May 10, 2025.

Tested with the built-in USB HD Webcam of my Panasonic CF-SV7.  Works in
chrome, cheese and with ffmpeg. Does not work in Firefox (waits for permissions)
or pwcview.

To record from webcam with ffmpeg:

    ffmpeg -f v4l2 -i /dev/video0 -vcodec rawvideo record.mkv

After closing the video device, you still have to kldunload uvc and
kldload uvc again, otherwise it either stops working or you'll get
strange artifacts.

* Applied some automatic DragonFly conversions using the following
  Makefile script:

    replace:
    	ruby -i -p -e '$$_.gsub!(%r[dev/usb], "bus/u4b")' *.c *.h
    	ruby -i -p -e '$$_.gsub!(%r[(malloc|free)\(], "k\\1(")' *.c *.h
    	ruby -i -p -e '$$_.gsub!(%r{([a-z]*)printf\(}, "k\\1printf(")' *.c *.h

* Add STAILQ_FOREACH_SAFE macros adapted from
  netproto/802_11/ieee80211_dragonfly.h

* Add atomic_subtract_64 macro

* Add kqueue support, replacing FreeBSD's poll code

* Fix mmap implementation

* Properly size uvc_data_request packages for UVC versions >= 1.5
  to 48 bytes (taken from OpenBSD). This is required to recognize
  my webcam.

[1]: https://github.com/lwhsu/freebsd-src/tree/feature-uvc-prepare-submit3