crypto - remove /dev/crypto pseudo-device
Remove the /dev/crypto pseudo-device. OpenBSD, which invented this API,
did this step back in release 5.7 (2015).

The last (and only) piece of software that made use of /dev/crypto in
DragonFly was tcplay and libtcplay (and cryptsetup), and the dependency
on /dev/crypto in these applications was replaced on April 21, 2025 by
doing the crypto operations synchronously in userland without any kernel
support.

Userland libraries or applications like OpenSSH or openssl should not
use /dev/crypto for performance reasons.