opencrypto - remove in-kernel crypto(9) framework
The opencrypto crypto(9) API was too complex (8k LoC) and unused
as of now.  It allowed for chaining various operations, crypto ops and
compression, but AFAIK this was never really used. For a much simpler,
synchronous API see sys/src/crypto/cryptoapi.

The opencrypto API was asynchronous by design. This was good back in the
days when dedicated hardware crypto devices did exist to help offload the
CPU. But the world has now changed towards synchronous CPU instructions
like AESNI. These dedicated CPU instructions are best called
synchronously, which removes the need for book-keeping of asynchronous
requests, often greatly simplifying the caller as well as the API and
"backend" implementation.

There is also no need to call those CPU instructions in-kernel, so
applications like OpenSSL can simply make use of them regardless of the
operating system.  This further avoids context switches going from
userland to kernel and vice versa.

For a much simpler and synchronous crypto API that both works in-kernel
as well as in userland (without the need to go through /dev/crypto
device), see src/sys/crypto/cryptoapi.